Fixed engine, vgui2, vguimatsurface
Also updated vpcs to include icommandline.h and strtools.h from the correct libs and thread jobs have been fixed

+++ src/engine/dt_send_eng.h
@@ -16,7 +16,7 @@
 #include "bitbuf.h"
 #include "utlmemory.h"
 
-typedef unsigned long CRC32_t;
+typedef unsigned int CRC32_t;
 
 class CStandardSendProxies;
